paper_search/requirements.txt
langgraph>=0.2.50
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
anthropic>=0.39.0
arxiv>=2.1.0
pydantic>=2.0
python-dotenv>=1.0.0
